refactor(auth): authId -> authCode, dto validation 추가, 패키지 분리
