+++ src/main/java/com/project/bumawiki/domain/auth/presentation/AuthController.java
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.project.bumawiki.domain.auth.presentation.dto.LoginRequestDto;
-import com.project.bumawiki.domain.auth.presentation.dto.RefreshTokenRequestDto;
-import com.project.bumawiki.domain.auth.presentation.dto.TokenResponseDto;
+import com.project.bumawiki.domain.auth.presentation.dto.request.LoginRequestDto;
+import com.project.bumawiki.domain.auth.presentation.dto.request.RefreshTokenRequestDto;
+import com.project.bumawiki.domain.auth.presentation.dto.response.TokenResponseDto;
 import com.project.bumawiki.domain.auth.service.CommandAuthService;
 
 import lombok.RequiredArgsConstructor;
@@ -25,7 +25,7 @@ public class AuthController {
 
 	@PostMapping("/oauth/bsm")
 	public TokenResponseDto userSignup(@RequestBody LoginRequestDto loginRequestDto) throws IOException {
-		return TokenResponseDto.from(commandAuthService.login(loginRequestDto.authId()));
+		return TokenResponseDto.from(commandAuthService.login(loginRequestDto.authCode()));
 	}
 
 	@PutMapping("/refresh/access")
