--- src/main/java/com/project/bumawiki/domain/auth/presentation/dto/LoginRequestDto.java
@@ -1,6 +1,0 @@
-package com.project.bumawiki.domain.auth.presentation.dto;
-
-public record LoginRequestDto(
-	String authId
-) {
-}
