--- src/main/java/com/project/bumawiki/domain/auth/presentation/dto/RefreshTokenRequestDto.java
@@ -1,6 +1,0 @@
-package com.project.bumawiki.domain.auth.presentation.dto;
-
-public record RefreshTokenRequestDto(
-	String refreshToken
-) {
-}
