+++ src/main/java/com/project/bumawiki/domain/auth/presentation/dto/request/LoginRequestDto.java
@@ -1,0 +1,9 @@
+package com.project.bumawiki.domain.auth.presentation.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record LoginRequestDto(
+	@NotNull(message = "authCode는 null일 수 없습니다.")
+	String authCode
+) {
+}
